requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
google-generativeai==0.8.3
pypdf==3.17.4
python-docx==1.1.0
numpy>=2.0.0
gunicorn==21.2.0